requirements.txt
langchain~=0.0.305
google-search-results
g4f==0.1.7.6
pydantic~=1.10.7
fastapi~=0.103.2
requests~=2.31.0
uvicorn~=0.23.2
starlette~=0.27.0
nest_asyncio
Pillow~=10.1.0